Move validation and constants into new files
Linting and cleanup
