+++ src/components/CodeAssignmentModal/CodeAssignmentModal.test.jsx
@@ -1,4 +1,5 @@
 /* eslint-disable react/prop-types */
+
 import React from 'react';
 import { Provider } from 'react-redux';
 
@@ -9,10 +10,8 @@ import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import { MemoryRouter } from 'react-router-dom';
 import remindEmailTemplate from './emailTemplate';
-import CodeAssignmentModal, {
-  getTooManyAssignmentsMessage, getInvalidEmailMessage, ASSIGNMENT_MODAL_FIELDS, BaseCodeAssignmentModal, getErrors,
-  textAreaKey, csvFileKey, NO_EMAIL_ADDRESS_ERROR, BOTH_TEXT_AREA_AND_CSV_ERROR,
-} from '.';
+import CodeAssignmentModal, { BaseCodeAssignmentModal } from '.';
+import { ASSIGNMENT_MODAL_FIELDS } from './constants';
 import { displayCode, displaySelectedCodes } from '../CodeModal/codeModalHelpers';
 
 import {
@@ -138,19 +137,22 @@ describe('CodeAssignmentModal component', () => {
   it.skip('displays an error', () => {
     jest.mock('redux-form', () => ({
       ...jest.requireActual('redux-form'),
+      // eslint-disable-next-line react/prop-types
       Field: ({ label, ...rest }) => <div {...rest}>{label}</div>,
     }));
     // eslint-disable-next-line global-require, no-unused-vars
     const { Field } = require('redux-form');
     const error = 'Errors ahoy!';
     const props = { ...initialProps, error: [error], submitFailed: true };
-    render(<MemoryRouter>
-      <Provider store={mockStore(initialState)}>
-        <BaseCodeAssignmentModal
-          {...props}
-        />
-      </Provider>
-    </MemoryRouter>);
+    render(
+      <MemoryRouter>
+        <Provider store={mockStore(initialState)}>
+          <BaseCodeAssignmentModal
+            {...props}
+          />
+        </Provider>
+      </MemoryRouter>,
+    );
 
     // const submitButton = screen.getByTestId('submit-button');
     // userEvent.click(submitButton);
@@ -172,169 +174,5 @@ describe('CodeAssignmentModal component', () => {
   it('renders a auto-reminder checkbox', () => {
     render(<CodeAssignmentModalWrapper />);
     expect(screen.getByText(ASSIGNMENT_MODAL_FIELDS['enable-nudge-emails'].label)).toBeInTheDocument();
-  });
-  describe('getTooManyAssignmentsMessage', () => {
-    it('displays the number of codes', () => {
-      const result = getTooManyAssignmentsMessage({
-        emails: ['foo@bar.com'],
-        numCodes: 10,
-        selected: false,
-      });
-      expect(result).toContain('You have 10');
-    });
-    it('pluralizes codes correctly', () => {
-      const resultPlural = getTooManyAssignmentsMessage({
-        emails: ['foo@bar.com'],
-        numCodes: 10,
-        selected: false,
-      });
-      const resultSingular = getTooManyAssignmentsMessage({
-        emails: ['foo@bar.com'],
-        numCodes: 1,
-        selected: false,
-      });
-      expect(resultPlural).toContain('codes ');
-      expect(resultSingular).toContain('code ');
-    });
-    it('shows selected or remaining codes', () => {
-      const resultRemaining = getTooManyAssignmentsMessage({
-        emails: ['foo@bar.com'],
-        numCodes: 10,
-        selected: false,
-      });
-      const resultSelected = getTooManyAssignmentsMessage({
-        emails: ['foo@bar.com'],
-        numCodes: 1,
-        selected: true,
-      });
-      expect(resultRemaining).toContain('remaining');
-      expect(resultSelected).toContain('selected');
-    });
-    it('gives the correct csv info', () => {
-      const resultCSV = getTooManyAssignmentsMessage({
-        emails: ['foo@bar.com'],
-        numCodes: 10,
-        selected: false,
-        isCsv: true,
-      });
-      const resultNoCSV = getTooManyAssignmentsMessage({
-        emails: ['foo@bar.com'],
-        numCodes: 10,
-        selected: false,
-      });
-      expect(resultCSV).toContain('your file has');
-      expect(resultNoCSV).toContain('you entered');
-    });
-    it('has the correct email length', () => {
-      const resultNoCSV = getTooManyAssignmentsMessage({
-        emails: ['foo@bar.com', 'bears@bearmountain.com'],
-        numCodes: 10,
-        selected: false,
-      });
-      expect(resultNoCSV).toContain('2 emails');
-    });
-  });
-  describe('getInvalidEmailMessage', () => {
-    it('returns a message with the invalid email address', () => {
-      const badEmail = 'bad@winnie.horse';
-      const result = getInvalidEmailMessage([2, 5], ['good@nico.horse', 'nice@donkey.com', badEmail, 'awesome@mule.com']);
-      expect(result).toContain(badEmail);
-    });
-    it('returns the line number of the email', () => {
-      const badIndex = 1;
-      const result = getInvalidEmailMessage([badIndex, 5], ['good@nico.horse', 'nice@donkey.com', 'awesome@mule.com']);
-      expect(result).toContain(`line ${badIndex + 1}`);
-    });
-  });
-  describe('getErrors', () => {
-    const sampleInputCsv = {
-      unassignedCodes: 1,
-      numberOfSelectedCodes: 1,
-      shouldValidateSelectedCodes: true,
-      invalidCsvEmails: [0],
-      validCsvEmails: ['another@me.com', 'yetOneMore@me.com'],
-      csvEmails: ['you@you.com', 'onemore@horse.com'],
-    };
-    const sampleInputTextArea = {
-      invalidTextAreaEmails: [1],
-      textAreaEmails: ['me@me.com', 'foo@bar.com'],
-      validTextAreaEmails: ['foo@bar.com', 'sue@bear.com'],
-      ...sampleInputCsv,
-      validCsvEmails: [],
-    };
-    it('returns an invalid email message for text area emails', () => {
-      const result = getErrors(sampleInputTextArea);
-      const errorMessage = getInvalidEmailMessage(sampleInputTextArea.invalidTextAreaEmails, sampleInputTextArea.textAreaEmails);
-      const expected = {
-        [textAreaKey]: errorMessage,
-        _error: [errorMessage],
-      };
-      expect(result).toEqual(expected);
-    });
-    it('returns an too many assignments message for the text area', () => {
-      const result = getErrors({ ...sampleInputTextArea, invalidTextAreaEmails: [] });
-      const errorMessage = getTooManyAssignmentsMessage({
-        emails: sampleInputTextArea.validTextAreaEmails, numCodes: sampleInputTextArea.unassignedCodes,
-      });
-      const expected = {
-        [textAreaKey]: errorMessage,
-        _error: [errorMessage],
-      };
-      expect(result).toEqual(expected);
-    });
-    it('returns too many assignments message when there are too many valid emails (text area)', () => {
-      const result = getErrors({ ...sampleInputTextArea, invalidTextAreaEmails: [], unassignedCodes: 2 });
-      const errorMessage = getTooManyAssignmentsMessage({
-        emails: sampleInputTextArea.validTextAreaEmails, numCodes: sampleInputTextArea.numberOfSelectedCodes, selected: true,
-      });
-      const expected = {
-        [textAreaKey]: errorMessage,
-        _error: [errorMessage],
-      };
-      expect(result).toEqual(expected);
-    });
-    it('returns an invalid email message if there are invalid CSV emails', () => {
-      const result = getErrors({ ...sampleInputCsv, unassignedCodes: 2 });
-      const errorMessage = getInvalidEmailMessage(sampleInputTextArea.invalidCsvEmails, sampleInputTextArea.csvEmails);
-      const expected = {
-        [csvFileKey]: errorMessage,
-        _error: [errorMessage],
-      };
-      expect(result).toEqual(expected);
-    });
-    it('returns too many assignments message for csv emails when there are more emails than codes', () => {
-      const result = getErrors({ ...sampleInputCsv, invalidCsvEmails: [] });
-      const errorMessage = getTooManyAssignmentsMessage({
-        isCsv: true,
-        emails: sampleInputCsv.validCsvEmails,
-        numCodes: sampleInputCsv.unassignedCodes,
-      });
-      const expected = {
-        [csvFileKey]: errorMessage,
-        _error: [errorMessage],
-      };
-      expect(result).toEqual(expected);
-    });
-    it('returns too many assignments message for csv when there are more emails than selected codes', () => {
-      const result = getErrors({
-        ...sampleInputCsv, unassignedCodes: 2, invalidCsvEmails: [], selectedCodes: 3,
-      });
-      const errorMessage = getTooManyAssignmentsMessage({
-        emails: sampleInputCsv.validCsvEmails, numCodes: sampleInputCsv.numberOfSelectedCodes, selected: true, isCsv: true,
-      });
-      const expected = {
-        [csvFileKey]: errorMessage,
-        _error: [errorMessage],
-      };
-      expect(result).toEqual(expected);
-    });
-    it('returns a no email addreses error if there are no email addresses', () => {
-      const result = getErrors({ validTextAreaEmails: [], validCsvEmails: [] });
-      expect(result).toEqual({ _error: [NO_EMAIL_ADDRESS_ERROR] });
-    });
-    it('returns an error if both text area and csv have emails', () => {
-      const result = getErrors({ validTextAreaEmails: ['foo'], validCsvEmails: ['bar'] });
-      expect(result).toEqual({ _error: [BOTH_TEXT_AREA_AND_CSV_ERROR] });
-    });
   });
 });
